fix: burnedCalories Double에서 Float로 변경
- 원래 burnedCalories가 Float이기 때문

+++ project/back/health/src/main/java/com/ssafy/health/domain/exercise/dto/response/ExerciseHistorySaveResponseDto.java
@@ -4,5 +4,5 @@ import lombok.Builder;
 
 @Builder
 public class ExerciseHistorySaveResponseDto {
-    private Double burnedCalories;
+    private Float burnedCalories;
 }
